- Implemented lazy-loading of resource libraries:
StreamFinder.GetFileStreamFromResources now checks a dictionary (should be a resource dictionary)
that contains the project name (key), and corresponding file name (value) of the libraries that
should be lazy-loaded.

The logic of loading the assembly is implemented in LibraryIO.EnsureDllIsLoaded, which operates
synchronously on both desktop and SL implementations.

My concern is regarding how "non-embedded" resources might work (SL default). Right now, all "resource"
files in SL are accessed via a different path structure than the embedded resources in the desktop
version. Possibly, we may want all lazy-loaded resources to be embedded (if this doesn't work).

Other concerns are related to unit-testing this process. Presumably, in SL the tests will have to start
wait for the hosting webserver. Testing on the desktop should not have these issues, but we would have to
specify a copy the .dll to be dumped in the bin folder (this is the "Copy To Output" property of a file
included in the project).