Building version of Vts/Vts.Desktop that initiates database refactoring.
Added Database<T> base class that abstracts away shared details for
reading and writing a file of "atomic" data units (like PhotonDataPoint).
So far, this has been implemented for a new CollisionInfo class. Currently, to
utilize this capability, one needs to implement

1) a data class to represent the "atomic" piece of data, to be streamed to/from
   file (see CollisionInfo, for example)
2) a serializer including read/write (see CollisionInfoSerializer, for example)
3) a database, which inherits from Database<T> (see CollisionInfoDatabase)
4) a database writer (see CollisionInfoDatabaseWriter, for example)

These four things could easily be consolidated to two or so classes, but for now
I'd like to keep it separated out based on a separation of concerns. Future work
may or may not use all of this infrastructure, so I want to be able to use one
piece at a time and see where each piece is being used as we build out the
framework.

Note that I haven't yet refactored PhotonDatabase to inherit from Database<T>.
*First*, we need to write tests around the basic functionality of the
CollisionInfo methods, and verify that they're reading/writing correctly. At that
point, we can bring PhotonDatabase in line.

Note the actual implementation of Database<T> is actually
Database<T,TSelfReferencingObject>. As is explained in the code, this is solely
to help the Database do more work. One practical benefit is that it removes the
requirement for a user of a derived class to cast their DataPoints object to
the correct type.

--HG--
branch : dc-110405-database_refactoring