Bug fix for serialization of Database<T>.
Detailed description from blog:

Carole Hayakawa - Apr 8, 2011 11:49 AM
David, thanks for all your hard work! I will look through your updates and familiarize
myself with the new code. I pulled the changes and reran Vts.Test. CollisionInfoDatabaseTests
run fine, but PhotonDatabaseTests errors at the end of "using", during the Dispose.
Dispose calls PostWriteAction which calls Database.WriteToXML, the stream is obtained
fine, but on the WriteObject get error: There was an error deserializing the object
of type Vts.MonteCarlo.SimulationInput. No set method for property '' in type ''.
Did you also get this error? If so, no problem, I will try to debug. If not, possibly
a file is missing from your push?

David Cuccia - Apr 8, 2011 11:58 AM
Here's the error: http://connect.microsoft.com/wcf/feedback/details/525090/wcf-serialization-error-message-does-not-specify-type-and-method
I was running the tests on the desktop, which have typically been more picky.
Those are both passing for me, but the SimulationInput serialization seems to be
a problem for SL. I'll check it out.

David Cuccia - Apr 8, 2011 12:25 PM
Error was a red herring...thanks to one of the comments on the page above, I was
able to deduce that it was the property *before* SimulationInput that was the problem,
not containing a "set" method. This was the DataPoints property of Databse<T>
in the base class. I've fixed (and am about to push) the following change with
IgnoreDataMemberAttribute:

[IgnoreDataMember]
public IEnumerable<T> DataPoints
{
get{ return _dataPoints; }
}

--HG--
branch : dc-110405-database_refactoring