Major redesign of VBs.  Used to take in detectorInputs and assign to VBs.  Now take in VBs in SimulationInput and with each their associated detectorInputs. Renamed TerminationDetector to SurfaceDetector and HistoryDetector to VolumeDetector.
SimulationInput now has an IList<IVirtualBoundaryGroup>
-IVirtualBoundaryGroup implementors are e.g. SurfaceBoundaryGroup,
GenericVolumeGroup.  This is like "VBinput" and will be serializable.
-IVirtualBoundaryGroups contain an IList<IDetectorInput> and a flag
designating whether to write to database or not.
-So the old SimulationInput design had an IList<IDetectorInput>,
here we've added a grouping layer to this list and a flag.  But this
simplifies much of our previous design because we don't have to
figure out which VB goes with which detector.  Its spelled out here.

In MCSimulation, this list of IVirtualBoundaryGroup gets filled
out into an IList<IVirtualBoundary>.
-IVirtualBoundary implementors are e.g. DiffuseReflectanceVB,
DiffuseTranmittanceVB, etc.

--Each IVirtualBoundary contains a IDetectorController which is
a Ilist<IDetector>.
--I have ISurfaceDetectorController and IVolumeDetectorController
because they implement different Tally methods and each inherit from
base IDetectorController.

---IDetector has children ISurfaceDetector and IVolumeDetector which
are the previous ITerminationDetector and IHistoryDetector, just
renamed.

Vts currently compiles, but Vts.Test does not.

--HG--
branch : ch-110322-transport_of_source_restructuring